+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
-- paw,ui: add initial support for item autofill
 - storage: allow to set a storage root with `PAW_HOME` env var
 - storage: add mock
+- paw,ui: add initial support for item autofill
+- paw,ui: add support to subtitle
+- ui: remove unused parameter
 
 ## 0.22.2 - 17 February 2024